좋아요기능 구현중 240110

+++ src/main/java/com/backend/jejutour_server/repository/TourRepository.java
@@ -7,7 +7,7 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface TourRepository extends JpaRepository<TourEntity,Integer>, TourRepositoryCustom {
+public interface TourRepository extends JpaRepository<TourEntity,Long>, TourRepositoryCustom {
 //    @Query("select m.name from TourEntity m")
 //    List<String> findNameList();
 List<TourEntity> findByItemsRegion2CdValue(Long region2cd_label);
